+++ example/shared-state-integration-vue/vue/package.json
@@ -9,7 +9,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@webview-bridge/vue": "workspace:^",
     "@webview-bridge/web": "workspace:^",
     "vue": "^3.5.12"
   },
