+++ example/shared-state-integration-vue/vue/src/bridge.ts
@@ -1,6 +1,8 @@
+import type { Bridge, BridgeStore } from "@webview-bridge/web";
+import { getCurrentInstance, onScopeDispose, ref } from "vue";
+
 import { linkBridge } from "@webview-bridge/web";
 import type { AppBridge } from "@webview-bridge-example-shared-state-integration-vue/react-native/types";
-import { ref } from "vue";
 
 export const isReady = ref(false);
 
@@ -13,3 +15,21 @@ export const bridge = linkBridge<AppBridge>({
 });
 
 export const bridgeStore = bridge.store;
+
+export function useBridge<T extends Bridge>(
+  store: Omit<BridgeStore<T>, "setState">,
+) {
+  const state = ref(store.getState());
+
+  const unsubscribe = store.subscribe((newState) => {
+    state.value = newState;
+  });
+
+  if (getCurrentInstance()) {
+    onScopeDispose(() => {
+      unsubscribe?.();
+    });
+  }
+
+  return state;
+}
