--- packages/vue/README.md
@@ -1,5 +1,0 @@
-# webview-bridge
-
-webview-bridge package for vue
-
-For more information, check out https://github.com/gronxb/webview-bridge
