--- packages/vue/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@webview-bridge/vue",
-  "type": "module",
-  "version": "1.6.2",
-  "description": "Fully Type-Safe Integration for React Native WebView and Web",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gronxb/webview-bridge.git"
-  },
-  "license": "MIT",
-  "files": [
-    "dist",
-    "package.json"
-  ],
-  "main": "dist/commonjs/index.cjs",
-  "module": "dist/module/index.mjs",
-  "types": "dist/typescript/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/typescript/index.d.ts",
-      "import": "./dist/module/index.mjs",
-      "require": "./dist/commonjs/index.cjs"
-    }
-  },
-  "scripts": {
-    "prepare": "pnpm build",
-    "build": "node esbuild.config.js && tspc --emitDeclarationOnly",
-    "test:type": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "@webview-bridge/web": "workspace:^",
-    "esbuild": "^0.19.4",
-    "vue": "^3.3.11"
-  },
-  "peerDependencies": {
-    "@webview-bridge/web": "*",
-    "vue": ">=2.7"
-  }
-}
