--- packages/vue/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ESNext",
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
-    "module": "ESNext",
-    "moduleResolution": "node",
-    "skipLibCheck": true,
-    "declaration": true,
-    "isolatedModules": true,
-    "jsx": "react-jsx",
-    "outDir": "dist/typescript",
-
-    /* Linting */
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "include": ["src"],
-  "extends": "../../tsconfig.base.json"
-}
